markitdown/requirements.txt
aspose-words
aspose-cells
pillow
pytesseract
